refactor: move some files and use "resolveBinding"
