+++ providers/MongodbProvider.ts
@@ -7,7 +7,7 @@ import {
 } from '@ioc:Zakodium/Mongodb/Odm';
 
 import { getMongodbModelAuthProvider } from '../src/Auth/MongodbModelAuthProvider';
-import { Database } from '../src/Database';
+import { Database } from '../src/Database/Database';
 import createMigration from '../src/Migration';
 import { BaseModel, BaseAutoIncrementModel } from '../src/Model/Model';
 import { field } from '../src/Odm/decorators';
@@ -15,21 +15,13 @@ import { field } from '../src/Odm/decorators';
 export default class MongodbProvider {
   public constructor(protected app: ApplicationContract) {}
 
-  private registerDatabase(): void {
-    this.app.container.singleton('Zakodium/Mongodb/Database', () => {
-      const { config, logger } = this.app;
-      return new Database(config.get('mongodb', {}), logger);
-    });
-  }
-
-  public register(): void {
-    this.registerDatabase();
+  private registerOdm(): void {
     this.app.container.singleton('Zakodium/Mongodb/Odm', () => {
       BaseModel.$setDatabase(
-        this.app.container.use('Zakodium/Mongodb/Database'),
+        this.app.container.resolveBinding('Zakodium/Mongodb/Database'),
       );
       BaseAutoIncrementModel.$setDatabase(
-        this.app.container.use('Zakodium/Mongodb/Database'),
+        this.app.container.resolveBinding('Zakodium/Mongodb/Database'),
       );
 
       return {
@@ -40,23 +32,40 @@ export default class MongodbProvider {
         field,
       };
     });
+  }
 
+  private registerDatabase(): void {
+    this.app.container.singleton('Zakodium/Mongodb/Database', () => {
+      const { config, logger } = this.app;
+      return new Database(config.get('mongodb', {}), logger);
+    });
+  }
+
+  private registerMigration(): void {
     this.app.container.singleton('Zakodium/Mongodb/Migration', () => {
       return createMigration(
-        this.app.container.use('Zakodium/Mongodb/Database'),
+        this.app.container.resolveBinding('Zakodium/Mongodb/Database'),
       );
     });
   }
 
+  public register(): void {
+    this.registerOdm();
+    this.registerDatabase();
+    this.registerMigration();
+  }
+
   public boot(): void {
     if (this.app.container.hasBinding('Adonis/Addons/Auth')) {
-      const Auth = this.app.container.use('Adonis/Addons/Auth');
+      const Auth = this.app.container.resolveBinding('Adonis/Addons/Auth');
       Auth.extend('provider', 'mongodb-model', getMongodbModelAuthProvider);
     }
   }
 
   public async shutdown(): Promise<void> {
-    const Database = this.app.container.use('Zakodium/Mongodb/Database');
+    const Database = this.app.container.resolveBinding(
+      'Zakodium/Mongodb/Database',
+    );
     return Database.manager.closeAll();
   }
 }
